Extract localize function
