+++ index.js
@@ -1,121 +1,6 @@
+const psuedoLocalizeString = require('./localize');
+
 const pseudoLocalization = (() => {
-  const ACCENTED_MAP = {
-    a: "ȧ",
-    A: "Ȧ",
-    b: "ƀ",
-    B: "Ɓ",
-    c: "ƈ",
-    C: "Ƈ",
-    d: "ḓ",
-    D: "Ḓ",
-    e: "ḗ",
-    E: "Ḗ",
-    f: "ƒ",
-    F: "Ƒ",
-    g: "ɠ",
-    G: "Ɠ",
-    h: "ħ",
-    H: "Ħ",
-    i: "ī",
-    I: "Ī",
-    j: "ĵ",
-    J: "Ĵ",
-    k: "ķ",
-    K: "Ķ",
-    l: "ŀ",
-    L: "Ŀ",
-    m: "ḿ",
-    M: "Ḿ",
-    n: "ƞ",
-    N: "Ƞ",
-    o: "ǿ",
-    O: "Ǿ",
-    p: "ƥ",
-    P: "Ƥ",
-    q: "ɋ",
-    Q: "Ɋ",
-    r: "ř",
-    R: "Ř",
-    s: "ş",
-    S: "Ş",
-    t: "ŧ",
-    T: "Ŧ",
-    v: "ṽ",
-    V: "Ṽ",
-    u: "ŭ",
-    U: "Ŭ",
-    w: "ẇ",
-    W: "Ẇ",
-    x: "ẋ",
-    X: "Ẋ",
-    y: "ẏ",
-    Y: "Ẏ",
-    z: "ẑ",
-    Z: "Ẑ"
-  };
-
-  const BIDI_MAP = {
-    a: "ɐ",
-    A: "∀",
-    b: "q",
-    B: "Ԑ",
-    c: "ɔ",
-    C: "Ↄ",
-    d: "p",
-    D: "ᗡ",
-    e: "ǝ",
-    E: "Ǝ",
-    f: "ɟ",
-    F: "Ⅎ",
-    g: "ƃ",
-    G: "⅁",
-    h: "ɥ",
-    H: "H",
-    i: "ı",
-    I: "I",
-    j: "ɾ",
-    J: "ſ",
-    k: "ʞ",
-    K: "Ӽ",
-    l: "ʅ",
-    L: "⅂",
-    m: "ɯ",
-    M: "W",
-    n: "u",
-    N: "N",
-    o: "o",
-    O: "O",
-    p: "d",
-    P: "Ԁ",
-    q: "b",
-    Q: "Ò",
-    r: "ɹ",
-    R: "ᴚ",
-    s: "s",
-    S: "S",
-    t: "ʇ",
-    T: "⊥",
-    u: "n",
-    U: "∩",
-    v: "ʌ",
-    V: "Ʌ",
-    w: "ʍ",
-    W: "M",
-    x: "x",
-    X: "X",
-    y: "ʎ",
-    Y: "⅄",
-    z: "z",
-    Z: "Z"
-  };
-
-  let opts = {
-    prefix: "",
-    postfix: "",
-    map: ACCENTED_MAP,
-    elongate: true
-  };
-
   const textNodesUnder = element => {
     const walker = document.createTreeWalker(
       element,
@@ -133,35 +18,10 @@ const pseudoLocalization = (() => {
     return textNodes;
   };
 
-  const psuedoLocalizeString = string => {
-    let pseudoLocalizedText = "";
-    for (let character of string) {
-      if (opts.map[character]) {
-        const cl = character.toLowerCase();
-        // duplicate "a", "e", "o" and "u" to emulate ~30% longer text
-        if (
-          opts.elongate &&
-          (cl === "a" || cl === "e" || cl === "o" || cl === "u")
-        ) {
-          pseudoLocalizedText += opts.map[character] + opts.map[character];
-        } else pseudoLocalizedText += opts.map[character];
-      } else pseudoLocalizedText += character;
-    }
-
-    // If this string is from the DOM, it should already contain the pre- and postfix.
-    if (
-      pseudoLocalizedText.startsWith(opts.prefix) &&
-      pseudoLocalizedText.endsWith(opts.postfix)
-    ) {
-      return pseudoLocalizedText;
-    }
-    return opts.prefix + pseudoLocalizedText + opts.postfix;
-  };
-
-  const pseudoLocalize = element => {
+  const pseudoLocalize = (element, options) => {
     const textNodesUnderElement = textNodesUnder(element);
     for (let textNode of textNodesUnderElement) {
-      textNode.nodeValue = psuedoLocalizeString(textNode.nodeValue);
+      textNode.nodeValue = psuedoLocalizeString(textNode.nodeValue, options);
     }
   };
 
@@ -197,17 +57,8 @@ const pseudoLocalization = (() => {
     subtree: true
   };
 
-  const start = (options = { strategy: "accented" }) => {
-    if (options.strategy === "bidi") {
-      // Surround words with Unicode formatting marks forcing
-      // the RTL directionality of characters.
-      opts.prefix = "\u202e";
-      opts.postfix = "\u202c";
-      opts.map = BIDI_MAP;
-      opts.elongate = false;
-    }
-
-    pseudoLocalize(document.body);
+  const start = (options) => {
+    pseudoLocalize(document.body, options);
     observer.observe(document.body, observerConfig);
   };
 
@@ -217,7 +68,8 @@ const pseudoLocalization = (() => {
 
   return {
     start,
-    stop
+    stop,
+    localize: psuedoLocalizeString,
   };
 })();
